Removing use of innerHTML, more comments to describe tests.
